feat: Se mejoraron los métodos para manejo de errores y se añadio listar_donaciones_button, obtener_donacion y mostrar_popup

+++ controllers/donaciones_controller.py
@@ -1,10 +1,13 @@
 import logging
 from models.donaciones import Donacion
-from models.salones import Salon
-from models.database import get_db
+from models.database import SessionLocal
 from sqlalchemy.orm import Session
 from sqlalchemy.exc import SQLAlchemyError
 from datetime import datetime
+from kivy.lang import Builder
+from kivy.uix.popup import Popup
+from kivy.uix.boxlayout import BoxLayout
+from components.styled_popup import StyledPopup
 
 # Configuración de logging
 logging.basicConfig(level=logging.INFO)
@@ -15,32 +18,40 @@ class DonacionesController:
         self.vista = vista
 
     def crear_donacion(self, cantidad, descripcion, unidad, equipo, fecha):
-        #Validacion de datos sencilla
+        """Crea una nueva donación en la base de datos."""
+        # Validación de datos
         if not cantidad or not descripcion or not unidad or not equipo or not fecha:
             self.vista.mostrar_error("Todos los campos son obligatorios.")
             return
-        db: SessionLocal()
+
+        try:
+            cantidad = float(cantidad)
+            fecha = datetime.strptime(fecha, '%Y-%m-%d').date()
+        except ValueError:
+            self.vista.mostrar_error("Formato de datos incorrecto. Verifique la cantidad y la fecha.")
+            return
+
+        db = None
         donacion_creada = False
         try:
+            db = SessionLocal()
             with db.begin():
-                fecha_obj = datetime.strptime(fecha, '%Y-%m-%d').date()
                 donacion = Donacion(
-                    cantidad=float(cantidad),
+                    cantidad=cantidad,
                     descripcion=descripcion,
                     unidad=unidad,
                     equipo=equipo,
-                    fecha=fecha_obj,
+                    fecha=fecha,
                 )
                 db.add(donacion)
                 logger.info(f"Donación creada: {donacion.id}")
                 donacion_creada = True
-        except ValueError:
-            self.vista.mostrar_error("Formato de fecha incorrecto. Debe ser YYYY-MM-DD.")
         except SQLAlchemyError as e:
             logger.error(f"Error al crear donación: {e}")
             self.vista.mostrar_error("Error al crear donación. Inténtalo de nuevo.")
         finally:
-            db.close()
+            if db:
+                db.close()
             if donacion_creada:
                 self.vista.mostrar_exito("Donación creada exitosamente.")
 
@@ -50,22 +61,18 @@ class DonacionesController:
             self.vista.mostrar_error("Todos los campos son obligatorios.")
             return
             
-        db: SessionLocal()
+        db = SessionLocal()
+        donacion_actualizada = False
         try:
             with db.begin():
                 donacion = db.query(Donacion).filter(Donacion.id == donacion_id).first()
                 if donacion:
-                    fecha_obj = datetime.strptime(fecha, '%Y-%m-%d').date()
                     donacion.cantidad = float(cantidad)
                     donacion.descripcion = descripcion
                     donacion.unidad = unidad
-                    donacion.fecha = fecha_obj
+                    donacion.fecha = datetime.strptime(fecha, '%Y-%m-%d').date()
                     donacion.equipo = equipo
-                    donacion.salones.clear()
-                    for salon_id in salones_ids:
-                        salon = db.query(Salon).filter(Salon.id == salon_id).first()
-                        if salon:
-                            donacion.salones.append(salon)
+                    donacion_actualizada = True
                     logger.info(f"Donación actualizada: {donacion.id}")
                 else:
                     self.vista.mostrar_error("Donación no encontrada.")
@@ -75,15 +82,25 @@ class DonacionesController:
             logger.error(f"Error al actualizar donación: {e}")
             self.vista.mostrar_error("Error al actualizar donación. Inténtalo de nuevo.")
         finally:
-            self.vista.listar_donaciones()
+            db.close()
+            logger.info("Conexión a la base de datos cerrada.")
+            if donacion_actualizada:
+                self.vista.mostrar_exito("Donación actualizada exitosamente.")
 
     def eliminar_donacion(self, donacion_id):
-        db: Session = next(get_db())
+        # Validación de datos
+        if not donacion_id:
+            self.vista.mostrar_error("ID de donación es obligatorio.")
+            return
+
+        db = SessionLocal()
+        donacion_eliminada = False
         try:
             with db.begin():
                 donacion = db.query(Donacion).filter(Donacion.id == donacion_id).first()
                 if donacion:
                     db.delete(donacion)
+                    donacion_eliminada = True
                     logger.info(f"Donación eliminada: {donacion_id}")
                 else:
                     self.vista.mostrar_error("Donación no encontrada.")
@@ -91,14 +108,88 @@ class DonacionesController:
             logger.error(f"Error al eliminar donación: {e}")
             self.vista.mostrar_error("Error al eliminar donación. Inténtalo de nuevo.")
         finally:
-            self.vista.listar_donaciones()
+            db.close()
+            logger.info("Conexión a la base de datos cerrada.")
+            if donacion_eliminada:
+                self.vista.mostrar_exito("Donación eliminada exitosamente.")
 
     def listar_donaciones(self):
-        db: Session = next(get_db())
+        """Obtiene la lista de donaciones desde la base de datos."""
+        db = None
         try:
+            db = SessionLocal()
             donaciones = db.query(Donacion).all()
-            self.vista.actualizar_lista_donaciones(donaciones)
-            logger.info("Donaciones listadas.")
+            logger.info(f"{len(donaciones)} donaciones obtenidas de la base de datos.")
+            if hasattr(self.vista, 'actualizar_lista_donaciones'):
+                self.vista.actualizar_lista_donaciones(donaciones)
+            else:
+                raise AttributeError("La vista no tiene el método 'actualizar_lista_donaciones'.")
+            return donaciones
         except SQLAlchemyError as e:
             logger.error(f"Error al listar donaciones: {e}")
-            self.vista.mostrar_error("Error al listar donaciones. Inténtalo de nuevo.")
+            if hasattr(self.vista, 'mostrar_error'):
+                self.vista.mostrar_error("Error al listar donaciones. Inténtalo de nuevo.")
+            return []
+        finally:
+            if db:
+                db.close()
+                logger.info("Conexión a la base de datos cerrada.")
+    
+    def listar_donaciones_button_handler(self):
+        """Método para manejar el evento de listar donaciones."""
+        self.listar_donaciones(self.vista)
+    
+    def obtener_donacion(self, id=None, fecha=None):
+        if not id and not fecha:
+            self.vista.mostrar_error("Debes proporcionar un ID o una fecha para obtener la donación.")
+            return None
+        db = SessionLocal()
+        try:
+            db.query(Donacion)
+            if id:
+                donacion = db.query(Donacion).filter(Donacion.id == id).first()
+            else:
+                donacion = db.query(Donacion).filter(Donacion.fecha == fecha).first()
+
+            if donacion:
+                logger.info(f"Donación encontrada: {donacion.id}")
+                self.mostrar_popup(f"Donación encontrada: {donacion.id}, {donacion.fecha}")
+                return donacion
+            else:
+                if id:
+                    logger.warning(f"Donación con ID {id} no encontrada.")
+                    self.vista.mostrar_error("Donación no encontrada.")
+                elif fecha:
+                    logger.warning(f"Donación con fecha {fecha} no encontrada.")
+                    self.vista.mostrar_error("Donación no encontrada.")
+                return None
+        except SQLAlchemyError as e:
+            logger.error(f"Error al obtener donación: {e}")
+            self.vista.mostrar_error("Error al obtener donación. Inténtalo de nuevo.")
+            return None
+        finally:
+            db.close()
+            logger.info("Conexión a la base de datos cerrada.")
+    
+    def mostrar_popup(self, mensaje, titulo="Información"):
+        """Muestra un popup con un mensaje."""
+        try:
+            popup_content = StyledPopup()
+            if isinstance(mensaje, dict):
+                # Si el mensaje es un diccionario, mostrar pares clave-valor
+                popup_content.set_content(mensaje)
+            else:
+                # Si el mensaje es un texto, mostrarlo directamente
+                popup_content.ids.popup_label.text = mensaje
+
+            popup = Popup(
+                title=titulo,
+                content=popup_content,
+                size_hint=(0.8, 0.4),
+                title_align="center",
+                title_color=(1, 1, 1, 1),  # Título en blanco
+            )
+            popup_content.ids.close_button.bind(on_release=popup.dismiss)  # Vincular el botón de cerrar
+            popup.open()
+        except Exception as e:
+            logger.error(f"Error al mostrar el popup: {e}")
